Add DNS domain as a full message, along with its own list type.
For now the message only includes a simple string, but may be
enhanced with additional attributes in the future.  These might
include whether aquilon dns hosts the domain, which servers provide
the domain, which locations use the domain, etc.

To preserve backwards compatibility, the host message still uses
a simple string for DNS domain.  If it needs this additional
information we'll have to add a new field.