added in parent and fullname for locations
In order to be able to trace up hubs for syslocs, etc, we need
to know the graph of how locations are connected. I've added
in the parents field into the protocol which allows for this.
For a particular node, you get name, fullname, location type
and parents. The parents is a list of shallowly populated location
nodes - each node only has the name and location type and does
NOT have the fullname or parents populated.