make two separate protocols to describe locations
The locations was misusing a protocol - parents were
typed as full location nodes, but were only partly
filled in. With this change we have a simple location
identity protocol, and a more fleshed out location
protocol. A location protocol contains a location
identity, and a list of location identities for the
parent. Also added optional row/columns.