add in clusters and resources
Fleshed out the protobufs for clusters and resources. Not sure
if it's documented anywhere, but I'm following the convention
that an object will contain just the name of another object
if that object could be a container for this, but will contain
the full object if the object is more granular. For example,
a cluster contains a fully fleshed out host for every host
within that cluster. However a host object contains only
the string name of the cluster to which it belongs