Include NetworkEnvironment in ServiceAddress
This adds NetworkEnvironment and DNSEnvironment types to the protocol
buffers and includes NetworkEnvironment in the ServiceAddress.  Whilst
it would have been more straightforward to have the network_environment
as a string, there is a long-term plan to have broker (user-facing)
output exclusively via protocol-buffers (avoiding duplicating what the
text formatters to at present).

Change-Id: Ifeca2c68cf359c8c65e7db4764866b305440a91e
Addresses-Issue: Jira/AQUILON-4164
Reviewed-By: Dave Reeve <Dave.Reeve@morganstanley.com>