aqdsystems: ServiceAddress: network_environment should not be required but optional
Fred Barnes added a required field for the ServiceAddress object
in commit 70361608. However, Google Protobuf's good practices are
to never add nor remove a required field when evolving a protocol,
in order to keep compatibility with previous versions (and thus
different versions interacting together).

This commit replaces this field from required to optional in order
to follow those good practices.