aqdparameters: Parameter: add 'holder' and 'parameter' details
The Parameter object only provided the path and value of a feature
parameter. This information contains the holder and the parameter
details, but needs parsing by the consuming process, which in turn
requires to know the list of holders for accurate results.
This patch allows the protobuf to contains directly this information,
hence making it easier on the consumer side.

Addresses-Issue: Jira/AQUILON-5776